+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+### Changed
+### Fixed
+
+## [1.8.0] - 2021-08-19
+
+### Added
+- Add levers to tune LB health check configuration [#950](https://github.com/PublicMapping/districtbuilder/pull/950)
+
+### Changed
+- Parameterize `max_old_space_size` [#942](https://github.com/PublicMapping/districtbuilder/pull/942)
+- Don't define container-level CPU reservation [#942](https://github.com/PublicMapping/districtbuilder/pull/942)
+- Remove CPU and memory limits for tasks on EC2 [#942](https://github.com/PublicMapping/districtbuilder/pull/942)
+- Add lever for tuning CPU allocation at the task level [#942](https://github.com/PublicMapping/districtbuilder/pull/942)
+- Scale `vm.max_map_count` with container memory usage [#942](https://github.com/PublicMapping/districtbuilder/pull/942)
+
+### Fixed
+- Fix CSV import by increasing max JSON payload to 25mb [#951](https://github.com/PublicMapping/districtbuilder/pull/951)
+- Fix CSV export for other user's projects [#943](https://github.com/PublicMapping/districtbuilder/pull/943)
+
 ## [1.7.2] - 2021-08-13
 
 ### Changed
@@ -284,7 +306,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.7.2...HEAD
+[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.8.0...HEAD
+[1.8.0]: https://github.com/publicmapping/districtbuilder/compare/1.7.2...1.8.0
 [1.7.2]: https://github.com/publicmapping/districtbuilder/compare/1.7.1...1.7.2
 [1.7.1]: https://github.com/publicmapping/districtbuilder/compare/1.7.0...1.7.1
 [1.7.0]: https://github.com/publicmapping/districtbuilder/compare/1.6.0...1.7.0